Added option to not subtract Mike's presynaptic spike time
Added a global option whether to subtract baseline at the presynaptic
spike time from traces. By not subtracting, the filtering becomes more
effective. Added a checkbox in the configuration window. Turning this
off and the filtering on results in about 20-30% reduction in STDs, but
doesn't affect amplitudes. There's still something wrong with the index
times output in the M file.